common/requirements.txt
# Copyright 2025 Google LLC.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

absl-py==2.1.0
google-auth==2.38.0
google-api-core==2.24.1
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.83.0
pytest==8.3.4
mimetype==0.1.5
fastapi==0.115.6
aiohttp==3.11.10
requests==2.32.3
pillow==11.1.0
google-cloud-firestore==2.20.0
opencv-contrib-python-headless==4.11.0.86
tadau==1.0.8
cloud-detect==0.0.13
tenacity==9.0.0